chore: mutate 함수 위치 이동
